agents/bq_optimization/user_agent/agent.py
import os
from dotenv import load_dotenv
from google.adk.agents import LlmAgent
from google.adk.agents.a2a import AgentClient # For A2A communication

# Load environment variables from .env file
load_dotenv()

LENS_AGENT_ADDRESS = os.getenv("LENS_AGENT_ADDRESS")

# Placeholder for A2A client to Agent 2
# This will be properly initialized when the agent is run,
# potentially in an __init__ or a setup function if we adopt a class structure,
# or used directly if functions for A2A calls are defined.
lens_agent_client = None
if LENS_AGENT_ADDRESS:
    # In a real scenario, AgentClient might require more parameters or setup
    # For now, this is a conceptual placeholder.
    # lens_agent_client = AgentClient(uri=LENS_AGENT_ADDRESS)
    pass # Actual initialization will depend on ADK patterns observed later

user_facing_agent = LlmAgent(
    name="BigQueryUserAgent",
    model="gemini-2.0-flash", # Or another suitable model
    description="Receives user requests for BigQuery optimization and interacts with a lens agent.",
    instruction="""
    You are a helpful assistant that receives requests from users regarding BigQuery optimization.
    Your primary role is to understand the user's query and then relay this information
    to a specialized BigQuery Lens Agent that will provide the actual optimization advice.

    When a user asks for BigQuery optimization:
    1. Acknowledge the request.
    2. Clearly state that you will consult the BigQuery Lens Agent.
    3. (Placeholder for A2A call) You will then send the user's query to the BigQuery Lens Agent.
    4. (Placeholder for A2A response) Once the Lens Agent responds, you will present its findings to the user.

    For now, if you receive a query, state that you've received it and would normally contact the Lens Agent.
    Example User Query: "Can you help me optimize this BigQuery query: SELECT * FROM my_table WHERE date > '2023-01-01'?"
    Your response (current): "I've received your BigQuery optimization request. I will consult the BigQuery Lens Agent for recommendations."
    """,
    # No tools for this agent directly, it delegates.
    tools=[]
)

# Example of how an A2A call might be conceptually structured
# async def ask_lens_agent(query: str):
#     if not lens_agent_client:
#         # This is a simplified approach. Robust error handling & initialization needed.
#         print("Lens Agent client not initialized. Make sure LENS_AGENT_ADDRESS is set.")
#         return "Error: Lens Agent not available."
#     try:
#         response = await lens_agent_client.send_message(message_content=query)
#         return response.content # Or however the response is structured
#     except Exception as e:
#         print(f"Error communicating with Lens Agent: {e}")
#         return f"Error: Could not get a response from Lens Agent: {e}"

# The root_agent is what the ADK framework will typically look for.
root_agent = user_facing_agent

# To make this agent discoverable or runnable via A2A (if it were to also *receive* A2A calls),
# it might need an A2A server setup, similar to other agents.
# For now, its primary role is to *make* A2A calls.
# If this agent itself needs to be hosted and called by an orchestrator via A2A,
# then an a2a_server.py or similar setup would be needed.
# Based on the plan, this agent is "front-facing", implying it might be called by an orchestrator
# or a platform component. Let's assume for now its A2A server setup will be similar to 'planner' or 'social' agents if needed.
